+++ .eslintrc.json
@@ -7,32 +7,30 @@
     "createDefaultProgram": true
   },
   "env": {
-    // 전역객체를 eslint가 인식하는 구간
-    "browser": true, // document나 window 인식되게 함
+    "browser": true,
     "node": true,
     "es6": true
   },
-  "ignorePatterns": ["node_modules/", "next.config.mjs"], // eslint 미적용될 폴더나 파일 명시
+  "ignorePatterns": ["node_modules/", "next.config.mjs"],
   "extends": [
     "airbnb",
     "airbnb-typescript",
     "airbnb/hooks",
     "next/core-web-vitals",
-    "plugin:@typescript-eslint/recommended", // ts 권장
-    "plugin:prettier/recommended", // eslint의 포매팅을 prettier로 사용.
-    "prettier" // eslint-config-prettier prettier와 중복된 eslint 규칙 제거
+    "plugin:@typescript-eslint/recommended",
+    "plugin:prettier/recommended",
+    "prettier"
   ],
   "rules": {
-    "react/react-in-jsx-scope": "off", // react 17부턴 import 안해도돼서 기능 끔
-    // 경고표시, 파일 확장자를 .ts나 .tsx 모두 허용함
+    "react/react-in-jsx-scope": "off",
     "react/jsx-filename-extension": ["warn", { "extensions": [".ts", ".tsx"] }],
-    "no-useless-catch": "off", // 불필요한 catch 못쓰게 하는 기능 끔
+    "no-useless-catch": "off",
     "react/jsx-props-no-spreading": "off",
     "jsx-a11y/label-has-associated-control": "off",
     "jsx-a11y/click-events-have-key-events": "off",
     "react/require-default-props": "off",
     "no-restricted-exports": "off",
     "import/no-extraneous-dependencies": "off",
-    "import/prefer-default-export": "off" // 배럴링 export/import에서 나는 에러 제거
+    "import/prefer-default-export": "off"
   }
 }
